refactor(net): move platform specific network code to a new `platform` module
